Continued small improvements and refactoring to Array and DeviceArray classes.
QKinect: #pragma-ed out warnings.
D3D11Utils: refactored rectangle utils into its own class.

+++ Kinect/src/QKinect.cpp
@@ -186,7 +186,7 @@ QKinect::QKinectEvent QKinect::poll( NUI_SKELETON_FRAME& skeleton, Image4ub& rgb
 	}
 
 	int eventIndex = waitMultipleResult - WAIT_OBJECT_0;
-	if( eventIndex >= nEvents )
+	if( eventIndex >= static_cast< int >( nEvents ) )
 	{
 		return QKinect_Event_Timeout;
 	}
